+++ src/PolySeries.jl
@@ -20,10 +20,11 @@ using EnzymeCore: within_autodiff
 import LinearAlgebra: mul!   # PolySeries.mul! extends LinearAlgebra.mul!
 include("polymap.jl")
 include("ctps.jl")
+include("composition.jl")
 export CTPS, PSDesc, pow, cst, element, findindex # assign!, reassign!
 export add!, addto!, sub!, subfrom!, scale!, scaleadd!, copy!, zero!, mul!, compose!, compose
 export set_descriptor!, get_descriptor, clear_descriptor!
-export CompositionWorkspace
+export CompositionWorkspace, CompositionPlan
 export PSWorkspace, borrow!, release!, decomposite
 export sin!, cos!, exp!, log!, sqrt!, sinh!, cosh!, asin!, acos!, sincos!, tan!, inv!, div!
 export pow!